+++ bin/ds_target_move.sh
@@ -6,7 +6,7 @@
 # Author.....: Stefan Oehrli (oes) stefan.oehrli@oradba.ch
 # Editor.....: Stefan Oehrli
 # Date.......: 2026.01.09
-# Version....: v0.5.3
+# Version....: v0.5.4
 # Purpose....: Move Oracle Data Safe targets and their referencing objects
 #              to another compartment for given target names/OCIDs.
 # Requires...: bash (>=4), oci, jq, lib/ds_lib.sh